Move organization and user models to concerns
